requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==4.1.0
boto3==1.26.150
botocore==1.29.150
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.6
Django==4.2
django-storages==1.13.2
ffmpeg-python==0.2.0
filelock==3.9.0
future==0.18.3
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.0
llvmlite==0.40.0
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper==20230314
pbr==5.11.1
platformdirs==3.0.0
prompt-toolkit==3.0.38
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.5.5
regex==2023.6.3
requests==2.31.0
s3transfer==0.6.1
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
sqlparse==0.4.4
stevedore==5.0.0
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.6
